Updated include dependencies and fixed default cell behavior
Updated the include dependencies in some of the
source files to just reuse what other includes
mention. I don't think this changes anything
in a practical sense, but I felt to do it anyways.

Also, the cell would always be enabled by default,
I added back in the functionality to disable cells
that are filled in by the puzzle on start.

Signed-off-by: mrandronov <mrandronov88@gmail.com>